+++ tests/TestHelper/Mvc/JsonTemplate.php
@@ -8,7 +8,7 @@
  * @copyright (c) 2018, Sebastian Rapetti
  * @license http://opensource.org/licenses/MIT MIT License
  */
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Linna\TestHelper\Mvc;